[server] Block OFFLINE->DROPPED ST to wait for drop partition action
Drop partition is performed asynchronously through consumer action in SIT. This behavior
allowed a race to occur when we have OFFLINE->DROPPED followed by OFFLINE->STANDBY for the
same partition on the same host. This is because we initialize the partition synchronously
in OFFLINE->STANDBY and there is a chance the previously queued drop partition consumer
action is executed after OFFLINE->STANDBY and that will cause ingestion to fail.

The fix is to block and wait on the drop partition consumer action future in the OFFLINE->DROPPED
state transition. This way when the host is picked again for OFFLINE->STANDBY for the same
partition, it can do so with a clean state without any lingering async actions.