[server][dvc] Increase graceful shutdown time for Da Vinci to reduce timeout exception (#1331)
For Da Vinci, when we close / retire a version, it will call shutdownIngestionTask API, which will try to unsubscribe and sync offset for graceful shutdown. Today the timeout is 10 seconds and it could timeout for topic with many partition assignment.
This will result in two kinds of exceptions:

When it is doing graceful shutdown and syncing offset, it may fail after the wait timeout as it will also remove storage engine
It may also failed to persist certain lingering data in drainer after timeout.
This PR tries to do minimal changes to reduce the noises in application that runs DVC by increasing the total timeout for per SIT shutdown wait.