[router][server] Reduce the granularity of latency histogram percentiles and disable key value profiling for single gets (#1377)
Currently, we record fine-grained percentiles for key and value of all single get requests. This bloats up the metrics significantly. This commit restricts key and value profiling only for cases where this config is enabled.

In addition to this, removed P77 and P90 metrics from latency metrics. We still have p50, p95, p99 and p99.9.