[server] Fix stats handling for SINGLE_GET requests (#1381)
In a previous commit, when removing some metrics, we ended up removing some key metrics about key and value size values for single_get requests as well. That change also led to NPEs as the caller expected the K/V stats to always be recorded for Single-get requests. This commit adds Avg and Max stats by default for key and value sizes in single get requests.