[server] Fix NPE in StoreAwarePartitionWiseKafkaConsumerService#handleUnsubscription (#1402)
This PR only fixes NPE exception in StoreAwarePartitionWiseKafkaConsumerService#handleUnsubscription. 2nd call to the same TP unsubscription will be ignored, as expected, as it should be decrease twice.

There is still race condition in upstream call where unsubscribe and unsubscribeAll from SIT/Participant thread can dual unsubscribe on the same TP, due to race condition in certain map bookkeeping operation. This PR does not fix that, due to other unclear performance concern. It is harmless for now, as 2nd call to the same TP will be ignored by consumer.