[vpj][controller] Do not optimize backup version deletion for TTL repush (#1325)
Today Venice deletes the repush source version as it is exactly same as repushed version immediately after the push. But for TTL repush since it differs from the source version this assumption is not correct. This PR prevents early deletion of TTL repush version and treats it as regular push.
Also it creates a new pushID for ttl repush which would help tracking regular repush vs ttl repush.