[build] Add checker framework for nullness type-checking (#1411)
We've been running into a lot of NPEs lately, so we decided to bring in the Checker
Framework for nullness type-checking to help catch and prevent these issues. The checker
is off by default and will be turned on gradually for each module. To enable it, pass the
`-PskipCheckerFramework=false` flag when running Gradle.

Example: `./gradlew -PskipCheckerFramework=false compileJava`

For more info about checker framework, see [Checker
Framework](https://checkerframework.org)