[controller][server] Remove SIT ready-to-serve check for A/A and non-AGG store during L/F transition (#1409)
Ready-to-serve check happens in two threads today:
(1) Drainer: which is reasonable
(2) SIT thread: there are two sub-cases:

During L/F transition: It is only for empty RT topic, and today we have HB for non-agg and A/A already, so it only applies to the lingering AGG system store. In this PR, it is checked against store data replication type, if it is AGG we will still have it. We will completely remove it once AGG mode is removed.
Before subscribe, it has a pre-check for short circuit to complete, and this is fine, as it is before the subscribe.
Also, I noticed that there are a bunch of integration tests which enables incremental push on NON-AA store. This is totally invalid setup and after removing the additional RTS check, these tests start to fail/very flaky.
To make sure we don't do this in test and prod, this PR added a new check in update store command in parent controller, which will fail loudly if the new update request is to set incremental push to be true on a non-A/A store.
I have another thinking about enabling all the configs that are enabled in production to be enabled in the test suite, but I'd like to do it in another separate PR, so it is easier for reviewers.

Beyonds that, fixed a flaky unit test that can throw NPE.