[test] Introduced NameRepository, StoreName and StoreVersionName
The new StoreName and StoreVersionName are meant to be used as map
keys for these types of entities, to provide a stronger type than
Strings. This is a similar idea to the PubSubTopic and its repo.

The NameRepository gives out shared (cached) instances. This, in
turn, provides efficiency benefits:

- Reduced garbage generation.
- Cached hash code.
- Equality checks (in the common case) can resolve via identity
  comparisons (though deep comparison also works correctly, in
  case the caching limit is reached).

In this commit, the new classes are only used in a unit test, and
they will be integrated later.