[duckdb] Utility to compare table schemas
This is used to check whether a pre-existing table on disk has a
schema equal to the one we want.

Did this by introducing various utility classes, including:

- ColumnDefinition
- TableDefinition
- IndexType
- SQLUtils

Also refactored unit tests to decouple from DuckDB, so that it's
easier to test other engines in the future (e.g., SQLite).