[admin-tool] Add a cluster batch processing framework command and a system store empty push task (#1254)
This is just a side-effect PR I created for batch processing all stores in cluster when I am empty pushing all system stores to apply config updates.
Add the new command so you can write your store-oriented task and execute it cluster-wide with the admin-tool. It supports basic checkpointing and parallel processing.
With this, there is a system store empty push task in this PR as well, it did sanity checks and empty push to all system stores for a specific user store.