[duckdb] Cleanup and refactor DVRT interface to be more configurable (#1445)
Previously, the user would have to define the key and output schema inside their DVRT implementation, which wasn't clean. Now, they're moved into the constructor. The user now only needs to provide the output schema, as the key schema will be automatically passed in by DVC.

I also added a constructor parameter for original output schema to support reader/writer schemas, but added a ToDo on how to implement the logic to pass in the correct schema.

Lastly, I make the DuckDB DVRT implementation more configurable, as it was previously more hardcoded.