[server][da-vinci] Instantiate the gracefulShutdownLatch lazily
Instantiate the gracefulShutdownLatch lazily to avoid waiting for the
latch when shutting down an ingestion task that was just intialized.
This is more of an issue in unit tests where we mock out certain parts and
the run() method of SIT never get to execute so waiting on the latch that was
originally initialized during instance instantiation will timeout.