[da-vinci][common] DVC consumer for materialized view (batch only)
Limiting the change to support batch only materialized view to keep the PR small.
Hybrid DVC consumer support and related features such as heartbeat  will be added in a separate PR

1. New DaVinciClientFactory APIs for creating DVC for a given view.

2. Defined a new "storeName" rule for views to be used for metrics reporting and shared DVC client.
See VeniceView.getStoreAndViewName for details. Trying to reuse much of the existing DVC code and
structure while providing the support for users to subscribe to the original store and different views
of the same store.

3. Introduced ReadOnlyMaterializedViewVersion and StoreViewBackend abstractions to minimize the
code change needed for DVC to understand and work with views. ReadOnlyMaterializedViewVersion
will provide view sepcific properties for a ReadOnlyVersion and StoreViewBackend will provide
view specific properties for a StoreBackend.