[dvc][duckdb] API change proposal to support DVRT checkpointing
This PR simply adds a getStorageMetadataService API to DVRT, with a default
implementation returning null. If the SIT gets null from DVRT, then nothing
changes compared to today. If it gets a non-null instance, then read/write
operations for OffsetRecord and StoreVersionState will go through the DVRT
SMS instance, rather than the regular one.

This will allow DVRT to implement checkpointing. In the case of DuckDB, it
means that the DuckDB file will be able to contain the checkpointing, and
the RocksDB metadata partition will be empty.