[controller] Dual write push job details to both parent and child region (#1477)
Add a dual write to child region for push job details, so when the system store is converted to A/A it will still be able to consume the updates.

The hard-coding of AGG mode in system store initialization routine will not be changed, as it will only kick in if no such store is found. We will update this after second phase when we remove the parent RT write.