[server] Fix Offset Lag Short Circuit (#1472)
This PR ensures that `LeaderFollowerStoreIngestionTask#reportIfCatchUpVersionTopicOffset()` does not prematurely measure and report offset lag by first checking that the latch was created in the first place. A latch is only created in `LeaderFollowerPartitionStateModel` when ingestion begins for a current version and not when the version is a future version. It previously only checked for `!isLatchReleased()`, which is true by default as the variable is initialized to `false`, even if the latch was not created.

More details in PR: !1472.