[fast-client] Added option to enable/disable retry budget
So far, the retry budget feature doesn't work well with low QPS
(retry requests are being rejected),
so this PR introduces a config option to enable/disable this feature
and it also disables this feature by default.
We will re-enable it by default once we figure out a way to make it
work for all kinds of scenarios.