[dvc][doc] Refactor DaVinciRecordTransformerConfig to use Builder pattern and update DVRT docs (#1501)
1. Previously, the configs for DaVinciRecordTransformerConfig were set by passing in all of the
values into its constructor. This means that anytime we wanted to add another config, it would be
a breaking change. To resolve this, I refactored DaVinciRecordTransformerConfig to use a Builder pattern.
2. Fixed a bug where if a user set storeRecordsInDaVinci, it would always bootstrap from VT and
ignore the on-disk state. This isn't deseriable behavior. To fix this, I introduced a new config
called alwaysBootstrapFromVersionTopic. Although most users won't use this, this can serve use-cases
where the user is only storing their records in-memory. It is more preferable for users to at least
have storeRecordsInDaVinci set to true so they can bootstrap from disk and utilize blob transfer.
3. If a user doesn't intend on modifying the schema of their records, the user won't need to pass
inoutputValueSchema and outputValueClass to the DaVinciRecordTransformerConfig anymore. This can
serve use cases where the user just wants to redirect their records to an alternative storage such as DuckDB.
4. Updated the DVRT docs to include the latest changes.