[server] Up-Convert PUT message schema in AAWC (#1504)
Today, in AAWC DCR code, we always use incoming schema as both reader/writer schema to deserialize the incoming PUT. The problem is, if we update the complex record in schema evolution by adding new nested field, and incoming PUT is using old schema, then after merge, the updated value record will be missing the newly added nested fields and unable to be serialized back by superset schema.

The fix is simple, just also use resolved schema as reader schema to up-convert the incoming PUT. The resolved schema is defined as: If old value schema and incoming value schema is the same, then it is the old value schema. Otherwise, use superset schema of them.

Behavior Change: This PR has certain behavior change: Due to PUT potentially being up-converted to superset schema, it may contain fields filled with default value that is not in the original PUT message. And it may override the existing field value.