[protocol] Add adminOperationProtocolVersion into proto schema AdminTopicMetadata (#1522)
We are introducing a new field to cluster's adminTopicMetadata, 
we will use this version as the SOT for admin operation (de)serializer.

In /getAdminTopicMetadata API response, we want to return version along with
other metadata inside the metadata map. 

Hence, we need to add the field in proto to set the value in response object.