[changelog] Fix metadata refresh deadlock scenario
There exists a deadlocking race condition when a version swap happens. The version swap data change listener locks itself
and then proceeds with the version swap on the consumer.  However, other function calls would also call refresh() on the
metadata repository.  This in turn would trigger more callbacks on the data change listener.  In some scenarios, this
can cause deadlocking.

This change makes sure that we no longer leverage manual refreshes and instead let the whole process get driven by the
scheduled metadata refresh.  This also makes the refresh time configurable, and seems to speed up tests (yay!) as there
were frequent retries on consuming events as we had to wait for the version swap to happen internally before we could
consume the events we expected.