[admin-tool] Modify the response of updateAdminTopicMetadata DIFFSIZEOVERRIDE (#1530)
This PR modifies the response of /updateAdminTopicMetadata to return more meaningful response.

Before, when we send request to update metadata within /AdminTopicMetadata or updating executionId
at store level, we return cluster name and store name (if applicable), and not returning the metadata itself.
With this PR, we now return applicable metadata within the endpoint response.