[admin-tool] Add stores set filter to cluster-batch-task (#1543)
* [admin-tool] Add store set filter to cluster-batch-task

Add new optional parameter STORE_FILTER_FILE to CLUSTER_BATCH_TASK to perform
cluster batch task only on the intersection of provided stores and stores in
the cluster. This is useful when you have a list of problematic stores that
you'd like to perform some repair on, e.g. re-push system store but instead
of doing it to all stores in a cluster you can just run CLUSTER_BATCH_TASK
with this store list for all clusters and it will only do what's necessary.

If needed in the future we can also add other set operations like complement.
Which will do some operation on the stores in the cluster excluding stores in
the provided filter file.