[compat] [controller] add a field `largestUsedRTVersionNumber` in store config (#1512)
When we delete and recreate a store, we do not want to use the same old real time topic name, because that might not have been fully cleaned up. We want to use a new name, and to create a new name, we need to know the previous name that was ever used. This config `largestUsedRTVersionNumber` will help in finding this out.