[changelog][client] Some clean up and bug fixes in CDC
A few fixes that include:

RMDUtils.hasOffsetAdvanced: This method did not handle correctly some cases where the offset vectors dimensionality didn't line up
VeniceAfterImageConsumer: Version swap could get stuck if the store repo for in the internal consumer and external consumer were
out of sync. This change adds a new method where the outer consumer can instruct the inner consumer directly which version it wants to
seek on.

Heartbeats in CDC: There was it seemed a forgotten and not so clever lambda attempt to record the timestamp.  This did some unfortunate math
and failed to perform a System.currentTimeMillis - timestamp, which in turn recorded a very large value. This fixes that and it now reports
the lag correctly.