[server] Use correct offset when non-agg leader is resubscribing for workload change action (#1547)
In Non-Aggregate mode, we reserve a special keyword for the upstream offset map.  
During the resubscribe-as-leader action, the current code does not look for the  
correct key in the map and resubscribes to the beginning of the RT topic during  
a pool jump. This results in a spike in leader HB delay and many duplicate  
messages (skipped).  

This PR fixes it by using the correct key for the upstream offset based on SIT  
mode.