[server] [client] [test] Global RT DIV: Chunking Support (#1385)
Added chunking support for DIV messages when they are produced to Kafka topics, as the size of the DIV message can surpass the `~1MB` Kafka message limit. The existing chunking mechanism for `PUT` messages is reused with a different message type `GlobalRtDiv`, including the `CHUNK` and `CHUNKED_VALUE_MANIFEST` values in the message's `schemaId`. ✂️