[server][vpj] Alter collection field delete to use schema defaults (#1557)
[server][vpj] Alter collection merge operations to use schema defaults when deleting

Prior to this, the FieldOpHandler was assuming an empty collection should be the result of clearing a field which is of a collection type.  However, this is contrary to the behavior of other types which instead set the field value to the default value specified in the schema.  This change normalizes that behavior.

Additionally, tweaked TTL repush to provide the incoming colo id as -1 so that should timestamp ties occurr, the TTL repush insert loses out to what's in the store.  This is basically a tweak of if the timestamp in ttl repush should be inclusive or not.  Setting to -1 seems to make this behavior consistent regardless of what colo this comes from (where previously it's possible for a particular colo to be zero indexed).