[changelog][client] Partial rollback of checkpoint maintenance
It's been reported that since patching a stricter checkpoint tracking on afterimageconsumer we've been filtering too many messages.

this doesn't effect the correctness of the consumer, but it does mean consumers will get played back more events then maybe strictly necessary.  However, as it stands this is causing some production issues, so this change just reverts that tweak back to what it was doing originally.