[controller] Allow AdminOperationSerializer to use lower version to serialize AdminOperation message (#1549)
This PR allows the AdminOperationSerializer to use any lower version to serialize any AdminOperation
message created by the latest protocol schema version.

This PR is part of the effort to remove the protol upgrade deployment order in AdminOperation. 
We will pull the protocol version from ZK as the SOT to serialize the message, and send the writer
version along with the message so that the reader can always deserialize the message.