[producer] Make ProducerTool prioritize newer schemas over older ones (#1585)
In "ProducerTool", the input is Avro's JSON encoded data, and we use Avro serialization test to see if the json encoded data maps to a particular Avro schema. This seems to not work as expected, and we end up picking an incorrect schema to write. As a short-term improvement, this commit changes the order in which schemas are tested for to be looked at from newest to oldest, as users are more likely to write data with newer schemas.