[dvc] Do not unsubscribe store from store repository during Da Vinci shutdown (#1594)
Da Vinci client throws a lot of exception during shutdown due to the fact    
that it is unsubscribing to store repo while doing other checkpointing in    
async way. They will throw exception when store is not found from the repo.    

There are two ways to fix it:    
(1) Just don't unsubscribe it from store repository, as the subscription is    
fixed during client life cycle and there is no harm to keep the existing    
subscription during shutdown    
(2) Make all the shutdown work first and then perform unsubscribe, which    
will have more changes and potentially slow down.    

This PR goes with (1) as it is simpler and cleaner to achieve.