[server][vpj][common][changelog] Chunking support for materialized view (#1566)
* [server][vpj][common] VPJ and backend chunking support for materialized view

1. Removed ComplexVeniceWriterAdapter and force CompositeVeniceWriter to use ComplexVeniceWriter
for child writers. This is because we need a new write API for writing to materialized view in
VPJ that will supply the partitions that it wrote to. This way VPJ and MaterializedViewWriter
are both writing to materialized view topic using the same APIs (CompositeVeniceWriter).

2. Expanded ComplexVeniceWriter#complexPut to take a nullable Consumer<int[]> partitionConsumer to
provide a partition array that it wrote to. This information is used to build the view partition
map that will be attached to the records in source region version topic.

3. The view partition map is attached as a PubSubMessageHeader to the records. Introduced new
PubSubMessageHeader: VENICE_VIEW_PARTITIONS_MAP_HEADER. Expanded LeaderMetadataWrapper to also
carry the view partition map. VeniceWriter will write records with VENICE_VIEW_PARTITIONS_MAP_HEADER
if the provided LeaderMetadataWrapper contains a view partition map. If the record is a large
record then all chunks and the manifest are all expected to carry this view partition map.

4. In L/F SIT we will only expect the view partition map in the PubSubMessageHeader during NR
pass-through. Ingestion exception will be thrown if corresponding partition info is not found
for a given materialized view. In addition, we only expect PUT messages during NR pass-through.

5. Added new method forwardPut in ComplexVeniceWriter to write records or chunks of records to
pre-determined materialized view topic partition(s) with new DIV but without any other processing.
This is like the materialized view equivalent of NR pass-through write. Forwarded records will
not contain view partition map. This is intentional so we only pay the overhead in NR source.

* Addressed comments and added chunk assembling support in CC consumer

* Fix tests

* Address comments and add unit tests

* Address comments and add an unit test