[server] Fix resubscription issue for partitions of stale store version to avoid useless ingestion.  (#1580)
When a store version is retired, the controller attempts to delete the store version from the ZooKeeper-based StoreRepository and sends a signal through the participant store to the server host to terminate the StoreIngestionTask (SIT). However, it is possible that the SIT thread, which checks the ingestion role logic, acts faster than the KILL SIT command taking effect on the server host.

Currently, store resubscription is triggered when a store version is deleted from the StoreRepository. This resubscription process can result in an ingestion role change. For example, if a store version is active-active enabled and later becomes stale and gets deleted, fetching this version from the StoreRepository will not provide information on whether it was active-active enabled. As a result, we assume it is active-active disabled, leading to an unintended ingestion role change.

Such resubscription triggers an unsubscribe and subscribe operation, which can occur before or after the SIT termination takes effect on removing all subscriptions of this SIT. As a result, some partitions may remain on the consumer.

For version topic partitions, the ConsumerSubscriptionCleaner can eventually remove them by checking if the version topic still exists. However, for leader real-time topic partitions of a stale version, the ConsumerSubscriptionCleaner is unable to clean them up.

Fixes Introduced in This PR:
Prevent unintended resubscription – Ensure that SIT does not trigger resubscription due to an ingestion role change when a stale store version is deleted from the StoreRepository and store is system store.

Co-authored-by: Hao Xu <xhao@xhao-mn3.linkedin.biz>