[controller] New Semantic Usage Detector (#1574)
This class will help detect the bad new semantic usage between two Avro schemas.

We consider an usage as bad if the object from current schema has type change or using non-default
value for a new field comparing to the target schema. It is important to detect since those fields will be
dropped during the serialization to lower version. 

If the data in those fields are important, we better fail the message fast.
The class will be used in AdminOperationSerializer to help detect bad message
(not ready to process) during deployment time.