[server] Fix multi-threading bug in shared multi-region PubSub configs (#1613)
In production, we found that servers had been querying PubSub offsets from an completely unintended region in some cases. This is pretty gnarly and can lead to data-loss.

We found a concurrency issue in "KafkaStoreIngestionService" where the shared server config object would be replaced by temporarily created configs for different regions. The expectation of this code was to only create this object for local use.