[server] Global RT DIV: Leader (#1576)
When the `isGlobalRtDivEnabled()` feature flag is enabled, the drainer will no longer sync
the VT DIV to the `OffsetRecord` in the `StorageEngine`. In the `ConsumptionTask` stage, leader
will produce a (Global) RT DIV to VT to replicate to the followers. On the callback, it will
enqueue a `DivSnapshot` (VT DIV + RT DIV) to the Drainer. Drainer will put the RT DIV
(possibly chunked) in the storage engine, and sync the OffsetRecord with VT DIV.
Note that the `GlobalRtDivState` will be compressed because it's being sent to VT.

The Drainer component will be handled in a future PR.