[common][router] Cache dimensions for otel  (#1532)
Problem:
Currently, every metric.record() call creates a new Attributes object with all the dimensions needed for that metric. As it also happens on the happy path and as Attributes creation does mergesort and deduplication, it will lead to high rate of object churn, potentially affecting GC and performance issues.

Solution:
This PR in its original form dealt with constructing a global cache with the key to the cache being the concatenation of all possible dimension values which proved to be inefficient as the new Attributes creation seemed to be faster than constructing the String key to the cache. So we decided to switch up the design to handle different cases separately in multiple sub classes of MetricEntityState and useEnumMap where ever possible (but with multiple levels based on how many dynamic dimensions are needed for that metric). Lookup in EnumMap happens to be a lot faster than HashMap with keys based on Object/String, especially trying to compute a composite String key on the fly.

There are multiple cases that needs to be handled:
case 1:  No dynamic dimensions: it should just use a precreated Attributes object which is based on the base dimensions which will always be the same for this instance.
case 2: Some dynamic dimensions with bound values which can be enumerated: we can consider precreating the Attributes for all combinations and fetch it during runtime. But that might be a huge memory overhead, especially if many of the combinations will never be used. so this PR relies on lazy initialization of Attributes for each set of dimensions in MetricEntityState whenever the first time record() is called with those dimensions resulting in creating and caching only the needed combinations. Note that, at least in the router context, storeName and clusterName which are unbounded strings falls under the base dimensions with how the existing stats objects are structured, so they remain the same for each RouterHttpRequestStats, so it doesn't fall under case 3.
case 3: Some dynamic dimensions which are unbounded: For instance some values that are String like upstream_service which identifies the caller. This PR doesn't have a sub class for that, but it fits in naturally with one another layer of HashMap on top of one of the case 2.

This PR creates multiple subclasses to MetricEntityState like below
1. MetricEntityStateBase (Base dimensions + 0 additonal dimensions)
2. MetricEntityStateOneEnum (Base dimensions + 1 additonal dimension which can be an Enum)
3. MetricEntityStateTwoEnums (Base dimensions + 2 additonal dimensions which can be Enums)
4. MetricEntityStateThreeEnums (Base dimensions + 3 additonal dimensions which can be Enums)

Note that the number of dimensions mentioned here are on top of some of the common/base dimensions which this instance would be associated with. For example, an instance of RouterHttpRequestStats#healthyRequestMetric will always have the same storeName, clusterName, and RequestType which will be considered base dimensions and it will have 3 additional dimenions like HttpResponseStatusCode, HttpResponseStatusCodeCategory, VeniceResponseStatusCodeCategory. This metric will now use MetricEntityStateThreeEnums.