[server] Deprecate OffsetRecord field leaderOffset and its usage to avoid ready-to-serve check using wrong offset; Fix bug that persist VT offset into upstream offset map (#1627)
No we should not use this OffsetRecord#leaderOffset anymore.    

This PR fixes two issues.    

The first issue is:    
This is legacy field and we already replaced it with upstreamOffsetMap.    
However, we are still updating this field from all the regions and our    
offset fetch logic can fall back can still hit this value in corner case.    

However, in the RTS check, when we are measuring offset lag for this region,    
we will eventually fall back to leaderOffset as persisted offset, if all the    
map in PCS does not contain region offset value. This is WRONG, because it is    
updated by all the regions. So essentially we are doing lag for    
region 3 = end offset of region 3 - persisted offset of region 1.    
And if Region 1 has more data and higher offset, this can be < 0 and mark it    
ready to serve before it is truly ready after consumer repair service kicks in.    

The second issue is:    
In non-AGG bug fix PR: #1547 , there is a regression introduced during refactor    
- previously we will only checkpoint calculated upstream offset map during    
TopicSwitch, not during any standby -> leader, but during the refactor we    
merged two methods, but the checkpoint should not happen unless we are    
consuming RTs. Actually, we should not checkpoint to OffsetRecord at all in    
the method, we should only do the in memory update and only sync to    
OffsetRecord during formal checkpoint time.    

The fix is just simply remove the line, and add a unit test to test against    
the behavior.