[vpj] Return Zstd configs after initialization (#1640)
In a previous PR, we made the "pushJobZstdConfig" field private, and it broke an internal consumer using this field. Ideally, we should not be relying on member variables to open access for some custom implementations.

This change makes the "initZstdConfig" function return the "pushJobZstdConfig" object so the caller has a handle on the object.

This change is backwards compatible because a void function now returns an object. Since it was previously "void", it is guaranteed that no one is expecting any value to be returned.